Fixed Image Location, fixed Textwrapping (simple)
1. Fixed Image Location
The Image in the PixDialog appeared in some strange
positions because of a bad line in the Sprite class.

2. Fixed Textwrapping (simple)
Implemented a simple textwrapping system. First of
it takes the whole string and divides it into substrings,
depending on \n's.
Also now the maximum number of lines are getting calculated,
and if the number of lines exceed that maximum, the box
gets higher to make all the lines fit in :)